Restructure repo - delete obsolete `docs` folder - move `notes` to `docs` - move all non-test code into `src`   - `client` -> `src/client`   - `cmd` -> `src/cmd`   - `backend/dblayer` -> `src/dblayer`   - `backend/server` -> `src/server`
