--- docs/artists.js
@@ -1,4 +1,0 @@
-async function getArtists() {
-  const resp = await fetch(SERVER_URL + '/artists')
-  document.getElementById("location").innerHTML = await resp.text()
-}
