--- docs/chords.js
@@ -1,3 +1,0 @@
-function getChords(artist, song) {
-  document.getElementById("location").innerHTML = `Showing chords for ${params.get("song")} by ${params.get("artist")}`
-}
