--- docs/index.html
@@ -1,71 +1,0 @@
-<!--
-  Jordy's Chordies - a web app for song chords
-      https://github.com/barrettj12/chords
-  Copyright 2022, Jordan Barrett (@barrettj12)
-      https://github.com/barrettj12
-  Licensed under the GNU AGPLv3.
-
-  docs/index.html
-  The homepage for the frontend, showing all artists in the DB.
--->
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Jordy's Chordies</title>
-
-   <!-- Stylesheets -->
-   <link href="https://unpkg.com/base-css-theme@1.1.3/base.css" rel="stylesheet">
-
-   <!-- Scripts -->
-   <script src="artists.js"></script>
-   <script src="songs.js"></script>
-   <script src="chords.js"></script>
-</head>
-<body>
-  <header>
-    <h1>Jordy's Chordies</h1>
-  </header>
-  <main>
-    <p id="serverinfo"></p>
-    <p id="location"></p>
-    <script>
-      let SERVER_URL;
-      // (async function() {
-      //   const resp = await fetch("server")
-      //   if (resp.ok) {
-      //     // Use local server with specified address
-      //     SERVER_URL = await resp.text()
-      //   } else {
-      //     // Use production server
-      //     SERVER_URL = "https://barrettj12-chords.herokuapp.com"
-      //   }
-      //   
-      // })()
-      fetch("server").then(
-        resp => resp.text().then(
-          txt => SERVER_URL = txt
-        ),
-        _ => SERVER_URL = "https://barrettj12-chords.herokuapp.com"
-      )
-      serverinfo.innerText = `Server: ${SERVER_URL}`
-
-
-      const params = new URLSearchParams(document.location.search);
-      if (!params.has("artist")) {
-        getArtists()
-      } else if (!params.has("song")) {
-        getSongs(params.get("artist"))
-      } else {
-        getChords(params.get("artist"), params.get("song"))
-      }
-    </script>
-  </main>
-  <footer>
-    jordan barrett, 2022
-    make donation
-    open source
-  </footer>
-</body>
-</html>
