--- docs/server
@@ -1,1 +1,0 @@
-http://localhost:8080
