--- docs/songs.js
@@ -1,3 +1,0 @@
-function getSongs(artist) {
-  document.getElementById("location").innerHTML = `Showing all songs for ${params.get("artist")}`
-}
